pyproject.toml
[project]
name = "knowledge-agents"
version = "0.1.0"
description = "Multi-agent system for knowledge retrieval (Confluence, SharePoint), automated ticketing (Jira), and AI-powered progress reporting using LangChain and LangGraph."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiodns>=3.4.0",
    "aiohttp>=3.11.18",
    "apscheduler>=3.11.0",
    "async-lru>=2.0.5",
    "fastapi>=0.115.12",
    "pydantic>=2.11.4",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.4.0",
    "pytz>=2025.2",
    "ruamel-yaml==0.17.32",
    "sqlalchemy[asyncio]==2.0.4",
    "sqlalchemy-json>=0.7.0",
    "uvicorn[standard]>=0.34.2",
    "uvloop>=0.21.0",
    "pymysql>=1.1.1",
    "asyncmy>=0.2.10",
    "greenlet>=3.2.2",
    "openai>=1.79.0",
    "jinja2>=3.1.6",
    "pytest-asyncio>=0.26.0",
    "pytest-datafiles>=3.0.0",
    "typer>=0.15.4",
    "python-dotenv>=1.1.0",
    "langchain-openai>=0.3.17",
    "pydantic-yaml>=1.4.0",
    "asyncpg>=0.30.0",
    "psycopg>=3.2.9",
    "psycopg-binary>=3.2.9",
    "langchain-postgres>=0.0.14",
    "html2text>=2025.4.15",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.4.5",
    "transformers>=4.51.3",
    "torch>=2.7.0",
    "langchain-huggingface>=0.2.0",
    "datasets>=3.6.0",
    "pytest-html>=4.1.1",
    "langchain-core>=0.3.61",
    "async-timeout==4.0.3",
    "langchain>=0.3.25",
    "langchain-experimental>=0.3.4",
    "ragas>=0.2.15",
    "asgi-lifespan>=2.1.0",
    "python-docx>=1.1.2",
    "markitdown[all,docx]>=0.0.2",
    "langgraph-supervisor>=0.0.27",
    "dateparser>=1.2.1",
    "pytablewriter>=1.2.1",
    "py-markdown-table>=1.3.0",
    "office365-rest-python-client>=2.6.2",
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
